Work around crashes caused by Godot overriding the class_name storage
If your export in compiled script mode crashes, you need to remove your
current user://overrides.cfg. THIS CHANGE IS ONLY EFFECTIVE ON SAVE!

To reiterate, this is caused because Godot simply dumps the
ProjectSettings when using save_custom. The project settings also store
the _global_script_classes that GDScript parses to find and determine
class_named scripts. On a further note: If modders choose to create new
scripts and reference them by name, it won't work because Godot is not
designed to support that because, again, these class_names are stored in
the project settings. A custom registration is needed.