Version Exporter: Ignore Mac OS key changes
Ignores these changes by parsing a regex expression defined in the
project settings. Whenever a key is changed to use a control or command
key, Godot will change the input map on mac. To fix the ignore, get the
changed diff from the CI logs, replace the file numbers with \d*,
escape the brackets, and paste the RegEx back into the ignore file.